:recycle: refactor: redis config 파일 통합
